Ensure Bubbles are reset on screen on
When you shut the screen off with a zoom level of >0 in a condition
where the zoom level should be reset on screen off (i.e. recents),
onZoomChanged sometimes waits right until screen on to reset the zoom
level back to 0, which means the Bubbles aren't in a reset position
anymore, since Bubbles are reset on screen off.